+++ src/Matrix.h
@@ -5,8 +5,8 @@
 #include <vector>
 
 
-static std::vector<int> OneHotEncode(int num, int size) {
-    std::vector<int> one_hot_vector;
+static std::vector<double> OneHotEncode(int num, int size) {
+    std::vector<double> one_hot_vector;
     one_hot_vector.resize(size);
     std::fill(one_hot_vector.begin(), one_hot_vector.end(), 0);
     one_hot_vector[num] = 1;
